Transform Mathematical into a class
