+++ lib/mathematical/corrections.rb
@@ -1,10 +1,10 @@
-module Mathematical
+class Mathematical
   module Corrections
     def apply_corrections(maths)
       adjust_lt_gt(maths)
     end
 
-    # from the itex website:
+    # from the itex website: http://bit.ly/1Et74ed
     # It is possible (though probably not recommended) to insert MathML markup
     # inside itex equations. So "<" and ">" are significant.
     # To obtain a less-than or greater-than sign, you should use \lt or \gt, respectively.
