+++ lib/mathematical/version.rb
@@ -1,3 +1,3 @@
-module Mathematical
+class Mathematical
   VERSION = '0.6.2'
 end
