+++ script/bench
@@ -19,5 +19,5 @@ puts "Size: " + ((file.bytesize * num) / 1000).to_s + " kilobytes\n"
 puts "Iterations: " + num.to_s + "\n"
 
 Benchmark.bm(40) do |x|
-  x.report("Rendering...                             ")  { for i in 1..num; Mathematical::Render.new.render(file); end }
+  x.report("Rendering...                             ")  { for i in 1..num; Mathematical.new.render(file); end }
 end
