+++ script/generate_samples
@@ -170,6 +170,6 @@ $$
 
 quality.each_pair do |filename, eq|
   filename = filename.gsub(/[\$\\]*/, '').gsub(/\s+/, '-').gsub(/[^a-zA-Z\d]/, '-').gsub(/-{2,}/, '-').gsub(/-$/, '').downcase
-  svg_content = Mathematical::Render.new(:base64 => false).render(eq)
+  svg_content = Mathematical.new(:base64 => false).render(eq)
   File.open("samples/quality/#{filename}.svg", "w") { |file| file.write svg_content["svg"] }
 end
