+++ test/mathematical/basic_test.rb
@@ -7,7 +7,7 @@ class Mathematical::BasicTest < Test::Unit::TestCase
   end
 
   def test_multiple_calls
-    render = Mathematical::Render.new
+    render = Mathematical.new
     render.render('$\pi$')
     output = render.render('$\pi$')['svg']
     assert_equal 1, output.scan(/<svg/).size, 'should only contain one svg'
