+++ test/mathematical/performance_test.rb
@@ -6,7 +6,7 @@ class Mathematical::BasicTest < Test::Unit::TestCase
     assert_nothing_raised do
       big_file = File.read('test/mathematical/fixtures/performance/big_file.text')
       speed = Benchmark.realtime do
-        MathToItex(big_file).convert { |equation| Mathematical::Render.new.render(equation) }
+        MathToItex(big_file).convert { |equation| Mathematical.new.render(equation) }
       end
 
       assert_operator speed, :<=, 5
